Issue 0000648
A nested transaction is occasional created under the right circumstances. This results in a query request to timeout after 60 seconds. If using pro, the UI will be stuck for 60 seconds during this timeout. This happens maybe 1 out of 4 times when continually deleting transforms.

I moved the nested transaction (refreshCache()) after the transaction.commit().