Issue 0000649
Enhanced the column match router expression to handle strings with multiple spaces.
Fixed a bug where if the value in the column match router ended in 'or', the router would not work correctly. 
Tick marks can be used to designate the value. If no tick marks are used, the value is still valid but all whitespace around the value string will be trimmed.