0005079: Tibero bulk loader failing because table and columns quoted in control file
It seems like the table can not be quoted, but the column names can be
quoted.
So, changing the code to quote the column names if
db.delimited.identifier.mode parameter is true, but not quote the table
name.
Also adding the hard coded strings OPTIONALLY ENCLOSED BY '"' and
ESCAPED BY '\\' to the control file.